requirements.txt
Flask~=3.1.1
Flask-Login~=0.6.3
Flask-SocketIO~=5.5.1
python-socketio~=5.13.0
reportlab~=4.4.2
Werkzeug~=3.1.3
numpy~=2.3.1
psycopg2-binary~=2.9.10
python-dotenv~=1.1.1
requests~=2.32.4